requirements.txt
pandas
mysql-connector-python
joblib
fastapi
uvicorn
streamlit
requests
scikit-learn
